refactor(test): re-organize unit tests for domain
