+++ oxytcmri/tests/unit/domain/mocks.py
@@ -7,17 +7,9 @@ from oxytcmri.domain.entities.mri import (
     VoxelData, AtlasSegmentation, DTIMap,
 )
 from oxytcmri.domain.ports.repositories import SubjectRepository, MRIExamRepository
-from oxytcmri.domain.use_cases.compute_dti_normative_values import (
-    ComputeDTINormativeValues,
-)
 import pytest
 from typing import List, Optional, Tuple
 
-
-# Entities
-
-
-# repositories
 @pytest.fixture
 def test_center():
     return Center(id=1, name="Test Center")
@@ -56,7 +48,6 @@ class MockInMemorySubjectRepository(SubjectRepository):
         ]
 
 
-# Mocks for tests
 class MockVoxelData(VoxelData[float]):
     """Mock for voxel data."""
 
@@ -126,49 +117,3 @@ class MockInMemoryMRIRepository(MRIExamRepository):
             subject_id=subject_id,
             data=self.dti_md_data + self.atlas_data,
         )
-
-
-class TestSubjectRepository:
-    def test_find_subjects_by_center(self, test_center):
-        # definitions
-        repository = MockInMemorySubjectRepository(test_center)
-
-        # execution
-        all_subjects = repository.find_subjects_by_center(test_center)
-        healthy_volunteers = repository.find_subjects_by_center(
-            test_center, subject_type=SubjectType.HEALTHY_VOLUNTEER
-        )
-        patients = repository.find_subjects_by_center(
-            test_center, subject_type=SubjectType.PATIENT
-        )
-
-        # assertions
-        assert len(all_subjects) == 3
-        assert len(healthy_volunteers) == 2
-        assert len(patients) == 1
-
-
-# Use cases
-class TestComputeDTIReferenceValues:
-    def test_execute_use_case(self, test_center):
-        # definitions
-        dti_metric = DTIMetric.MD
-        atlas = Atlas(
-            id=2,
-            labels=[1, 2, 3],
-            name="Neuromorphometrics atlas + GM parcels size ≤5cm3",
-        )
-
-        # execution
-        use_case = ComputeDTINormativeValues(
-            subjects_repository=MockInMemorySubjectRepository(test_center),
-            mri_repository=MockInMemoryMRIRepository(atlases=[atlas]),
-        )
-        result = use_case.execute(test_center, dti_metric, atlas)
-
-        # assertions
-        assert result is not None
-        assert all(item.center == test_center for item in result)
-        assert all(item.dti_metric == dti_metric for item in result)
-        assert all(item.atlas == atlas for item in result)
-        assert all(isinstance(item.value, float) for item in result)
